backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.8
google-generativeai==0.3.2
chromadb==0.4.18
gitpython==3.1.40
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
sentence-transformers==2.2.2
huggingface-hub==0.19.4